refactor: expose unified qt resolver
